requirements.txt
# RA_Agent Requirements
# Document Analysis and Report Generation Pipeline with LangChain

# Core Dependencies
pdf2image>=1.16.0
pillow>=10.0.0
python-dotenv>=1.0.0

# LangChain Core & Extensions
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langgraph>=1.0.0
langgraph-checkpoint>=4.0.0
langsmith>=0.6.0

# LangChain PostgreSQL / PGVector
langchain-postgres>=0.0.6
psycopg2-binary>=2.9.9
pgvector>=0.2.0
sqlalchemy>=2.0.0

# AI API Clients
openai>=1.12.0
google-generativeai>=0.8.0
google-genai>=1.60.0
google-ai-generativelanguage>=0.6.0
google-api-core>=2.20.0
google-auth>=2.48.0
protobuf>=4.25.0,<5.0.0

# HTTP & Async
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.28.0
httpcore>=1.0.0

# OCR Services
pytesseract>=0.3.10

# PDF Processing
PyMuPDF>=1.23.0

# Data Processing
numpy>=1.24.0
pydantic>=2.0.0
PyYAML>=6.0.0

# Utilities
chardet>=5.2.0
tenacity>=9.1.0
tiktoken>=0.12.0

